+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtual/index.js
@@ -1,5 +1,6 @@
 import { useLayoutEffect, useEffect } from "react";
 import useVirtualModel from "../useVirtualModel";
+import { DEFAULT_OVERSCAN_COUNT } from "constants";
 
 const localUseEffect = process.env.__IS_SERVER__ ? useEffect : useLayoutEffect;
 
@@ -8,7 +9,8 @@ const useVirtual = params => {
 
     localUseEffect(() => {
         model._startBatch();
-        model.setHorizontal(params.horizontal);
+        model.setOverscan(params.overscanCount ?? DEFAULT_OVERSCAN_COUNT);
+        model.setHorizontal(!!params.horizontal);
         model.setItemCount(params.itemCount, params.getEstimatedItemSize);
         model._endBatch();
     });