+++ packages/af-utils-react-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-table
 
+## 0.0.36
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.31
+    -   @af-utils/styled@0.0.13
+
 ## 0.0.35
 
 ### Patch Changes