+++ packages/af-utils-styled/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/styled
 
+## 0.0.13
+
+### Patch Changes
+
+-   minor hooks refactor
+
 ## 0.0.12
 
 ### Patch Changes